Made the cd builtin
